Move classification into the console and away from trusted facts
Prior to this commit, we had users create a trusted fact that
sort of magically ended up classifying their PE master with the role
that we wanted.

After this commit, we will instruct users to add the appropriate
role to their PE master node via the console UI.  This simplifies
the process of classifying the node if you already have PE installed.